Backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
azure-cognitiveservices-speech==1.37.0
azure-core==1.30.1
azure-cosmos==4.6.0
azure-functions==1.18.0
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
fastapi==0.110.1
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
isodate==0.6.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
python-multipart==0.0.9
requests==2.31.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn>=0.11.0
azure-appconfiguration==1.6.0
